langsmith_demo/pyproject.toml
[project]
name = "langsmith-demo"
version = "0.1.0"
description = "Demos for pydantic-ai TemporalAgent integration with LangSmith tracing"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai>=1.40.0",
    "temporalio>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
    "langsmith>=0.6.1",
    "langchain-openai>=0.3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
